# sandboxd-orch/docs/orchestrator_swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Orchestrator API Server",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/sandboxes": {
            "get": {
                "description": "Lists orchestrator sandbox objects and their scheduling status.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator-sandbox"
                ],
                "summary": "List control-plane sandboxes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SandboxObjectsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a scheduler-managed sandbox object in orchestrator.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator-sandbox"
                ],
                "summary": "Create control-plane sandbox object",
                "parameters": [
                    {
                        "description": "Control-plane sandbox request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.CreateSandboxObjectRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.SandboxObjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/sandboxes/{id}": {
            "get": {
                "description": "Returns one orchestrator sandbox object.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator-sandbox"
                ],
                "summary": "Get control-plane sandbox",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SandboxObjectResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes orchestrator sandbox object and requests runtime deletion on assigned node.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orchestrator-sandbox"
                ],
                "summary": "Delete control-plane sandbox",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sandbox ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DeleteSandboxResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.DeleteSandboxResponse": {
            "type": "object",
            "properties": {
                "deleted": {
                    "type": "string"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.SandboxObjectResponse": {
            "type": "object",
            "properties": {
                "sandbox": {
                    "$ref": "#/definitions/types.Sandbox"
                }
            }
        },
        "handlers.SandboxObjectsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.Sandbox"
                    }
                }
            }
        },
...
        "types.CreateSandboxObjectRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "spec": {
                    "$ref": "#/definitions/types.SandboxSpec"
                }
            }
        },
...
        "types.ReadinessProbeSpec": {
            "type": "object",
            "properties": {
                "failure_threshold": {
                    "type": "integer"
                },
                "initial_delay_seconds": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "period_seconds": {
                    "type": "integer"
                },
                "port": {
                    "type": "integer"
                },
                "protocol": {
                    "type": "string"
                },
                "success_threshold": {
                    "type": "integer"
                },
                "timeout_seconds": {
                    "type": "integer"
                }
            }
        },
        "types.Sandbox": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "spec": {
                    "$ref": "#/definitions/types.SandboxSpec"
                },
                "status": {
                    "$ref": "#/definitions/types.SandboxStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "types.SandboxContainerSpec": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "env": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "resource": {
                    "$ref": "#/definitions/types.SandboxResource"
                },
                "work_dir": {
                    "type": "string"
                }
            }
        },
        "types.SandboxPhase": {
            "type": "string",
            "enum": [
                "Pending",
                "Scheduled",
                "Running",
                "Failed",
                "Deleting"
            ],
            "x-enum-varnames": [
                "SandboxPhasePending",
                "SandboxPhaseScheduled",
                "SandboxPhaseRunning",
                "SandboxPhaseFailed",
                "SandboxPhaseDeleting"
            ]
        },
        "types.SandboxPortAssign": {
            "type": "object",
            "properties": {
                "container_port": {
                    "type": "integer"
                },
                "host_port": {
                    "type": "integer"
                },
                "protocol": {
                    "type": "string"
                }
            }
        },
        "types.SandboxPortSpec": {
            "type": "object",
            "properties": {
                "container_port": {
                    "type": "integer"
                },
                "host_port": {
                    "type": "integer"
                },
                "protocol": {
                    "type": "string"
                }
            }
        },
        "types.SandboxResource": {
            "type": "object",
            "properties": {
                "cpu": {
                    "type": "string"
                },
                "ephemeral_storage": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                }
            }
        },
        "types.SandboxSpec": {
            "type": "object",
            "properties": {
                "containers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SandboxContainerSpec"
                    }
                },
                "egress": {
                    "type": "boolean"
                },
                "ports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SandboxPortSpec"
                    }
                },
                "readiness_probe": {
                    "$ref": "#/definitions/types.ReadinessProbeSpec"
                },
                "ttl_seconds": {
                    "type": "integer"
                }
            }
        },
        "types.SandboxStatus": {
            "type": "object",
            "properties": {
                "assigned_ports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.SandboxPortAssign"
                    }
                },
                "expire_at": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "last_error": {
                    "type": "string"
                },
                "node_name": {
                    "type": "string"
                },
                "phase": {
                    "$ref": "#/definitions/types.SandboxPhase"
                },
                "sandboxd_id": {
                    "type": "string"
                }
            }
        }
    }
}